third_party/blink/web_tests/wpt_internal/webgpu/canvas_webgpu_transfer/canvas-reset-destroys-texture.https.html

<!DOCTYPE html>
<title>Canvas2D WebGPU Transfer: canvas-reset-destroys-texture</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="./webgpu-helpers.js"></script>
<script src="./canvas-test.js"></script>

<h1>Canvas2D WebGPU Transfer: canvas-reset-destroys-texture</h1>
<p>Resetting a canvas during WebGPU access should destroy the GPUTexture.</p>

<script>
  runCanvasTestsInWorker({dependencies: ['./webgpu-helpers.js']});

  for (const [description, resetCanvas] of [
    ['Calling context.reset()', (ctx) => ctx.reset()],
    ['Resizing the canvas', (ctx) => ctx.canvas.width = ctx.canvas.width],
  ]) {
    canvasPromiseTest(async (canvas) => {
      // Skip this test on Mac Swiftshader due to "Invalid Texture" errors.
      const adapter = await getWebGPUAdapter();
      if (isMacSwiftShader(adapter.info)) {
        return;
      }

      const device = await getWebGPUDevice(adapter);

      // Begin a WebGPU access session.
      const ctx = canvas.getContext('2d');
      const tex = ctx.transferToGPUTexture({device: device,
                                            usage: GPUTextureUsage.COPY_SRC});

      // Reset the canvas. This should abort the WebGPU access session.
      resetCanvas(ctx);

      // The canvas' GPUTexture should appear to be destroyed.
      assert_true(await isTextureDestroyed(device, tex));
    }, description + ' during WebGPU access destroys the GPUTexture.');
  }
</script>
